Check in initial automation scripts for jenkins
This script sets up the directories needed for the build (build\build-windows-x64-debug, for this run), runs gen-buildsys-win.bat (or –clang, for linux), runs msbuild/make to build, then cleans up after itself by removing the build directories.  It has a stub for copying binaries to a remote directory once we have that set up.